got working aes128 implementation
git-svn-id: svn+q:///qpdf/trunk@809 71b93d88-0707-0410-a8cf-f5a4172ac649

+++ libqpdf/rijndael.cc
@@ -1,8 +1,8 @@
 #define FULL_UNROLL
 
-#include "rijndael.h"
+#include "qpdf/rijndael.h"
 
-typedef unsigned long u32;
+typedef uint32_t u32;
 typedef unsigned char u8;
 
 static const u32 Te0[256] =
